+++ system/lib/wasmfs/support.h
@@ -5,12 +5,14 @@
 
 #pragma once
 
-#include <cstdlib>
+#include <stdnoreturn.h>
 
+#include <cstdlib>
+ 
 #ifndef NDEBUG
 // In debug builds show a message.
 namespace wasmfs {
-void handle_unreachable(const char* msg, const char* file, unsigned line);
+noreturn void handle_unreachable(const char* msg, const char* file, unsigned line);
 }
 #define WASMFS_UNREACHABLE(msg)                                                \
   wasmfs::handle_unreachable(msg, __FILE__, __LINE__)
